Change test, extend config, other cleanup

+++ aintelope/environments/savanna.py
@@ -110,9 +110,9 @@ class HumanRenderState:
 def reward_agent(min_grass_distance):
     # For now measure if agent can eat. Was #1 / (1 + min_grass_distance), moving to instincts
     if min_grass_distance > 1:
-        return 0
+        return np.float64(0.0)
     else:
-        return 1
+        return np.float64(1.0)
 
 
 def move_agent(
